// src/demo/meetingsdk/types.ts
export type AccountConfig = {
  sdkKey: string;
  sdkSecret: string;
  webEndpoint: string;
  account: string;
  env: string;
};

export type FormData = {
  sdkKey: string;
  sdkSecret?: string;
  displayName: string;
  email?: string;
  meetingNumber: string;
  password: string;
  version: { value: string; label: string };
  role: { value: string; label: string };
  language: { value: string; label: string };
  web: { value: string; label: string };
  debug: { value: string; label: string };
  useVideoPlayer: { value: string; label: string };
  customerKey: string;
  enforceWebRtcAudio: { value: string; label: string };
  enforceWebRtcVideo: { value: string; label: string };
  geoRegions?: string[];
  telemetryId?: string;
  mediaVersion?: string;
  account: { value: string; label: string };
  cdn: { value: string; label: string };
  corp: { value: string; label: string };
  enforceGalleryView: { value: string; label: string };
  enforceVirtualBackground: { value: string; label: string };
  signatureToken?: string;
  widget: { value: string; label: string };
  china: { value: string; label: string };
  report: { value: string; label: string };
  waitingRoomPreview: { value: string; label: string };
  meetingPreview: { value: string; label: string };
  hd: { value: string; label: string };
  webcodecs: { value: string; label: string };
  sab: { value: string; label: string };
  externalLink: { value: string; label: string };
  iframeSize: { value: string; label: string };
  defaultView: { value: string; label: string };
  shareAudio: { value: string; label: string };
  test: { value: string; label: string };
  customizeVB: { value: string; label: string };
  customizeJoin: { value: string; label: string };
  pip: { value: string; label: string };
  invitePhone: { value: string; label: string };
  enableMobileUI: { value: string; label: string };
  zoomLogo: { value: string; label: string };
  leaveOnPageUnload: { value: string; label: string };
  autoAdmit: { value: string; label: string };
  customizedLeave: { value: string; label: string };
  exitRoomCallback: { value: string; label: string };
  leaveRedirect: { value: string; label: string };
  recordingToken?: string;
  buildVersion?: string;
  accountTk?: string;
  accountJmak?: string;
  meetingSignature?: string;
  meetingZak?: string;
  obfToken?: string;
  childToken?: string;
  rwc?: { value: string; label: string };
  build?: any;
  enforcePEPC: { value: string; label: string };
  selectedEvents: { value: string; label: string }[];
};

export type InputConfig = {
  name: string;
  placeholder: string;
  icon: React.ReactNode;
  readOnly?: boolean;
};